Modified version of E. Izquierdo's CTRNN class
In addition to interneuron weights, this class also includes an InputWeight vector, which govern the sensitivity of each neuron to some external input. This is set by the genome like the rest of the parameters. The dimensionality of the input can be up to the number of neurons in the network.