Added setInitialState() method for initializing state, and crucially, optical information
This method takes the initial state of the system (velocity, distance, target_size) and simulates a few steps of constant motion in order to give the optical variables a chance to stabilize. This is necessary because the some of the variables are the first or second derivative of the image size. It takes three time steps for these values to stabilize (confirmed via testing).