Added showTrajectory() method, removed data attributes
Justification: the only reason that I would need to access this fine-grained level of data is if I'm trying to understand one particular trajectory. But because this system is completely deterministic, I can re-create that trajectory reliably simply with access to the starting conditions.

The only exception to this is the brake mapping history, which could be modified over the duration of the trial. This continues to be stored.