Major bugfix, addition of diversity measures
The way that fitness was being recorded and reported was somehow different despite using the same underlying method to do each. I did not find the bug after several hours, so I've simply rolled back the changes to an earlier version of the mga. This means that I lose the individual fitness trajectory functionality.
I then added some measures for gene pool diversity and convergence, which are now also reported by fitStats().